Move importer/common.js file to state/imports
